Remove passthrough exports on ingest/granule.js.
After extracting cmr code from ingest/granule.js remove the explicit exports
and have tasks point to the extracted code.
