+++ packages/ingest/granule.js
@@ -24,8 +24,6 @@ const { httpMixin } = require('./http');
 const { s3Mixin } = require('./s3');
 const { baseProtocol } = require('./protocol');
 const {
-  getGranuleId,
-  getCmrFiles,
   updateMetadata
 } = require('./granule-extract-cmr');
 
@@ -848,8 +846,6 @@ module.exports.S3Granule = S3Granule;
 module.exports.S3DiscoverGranules = S3DiscoverGranules;
 module.exports.SftpDiscoverGranules = SftpDiscoverGranules;
 module.exports.SftpGranule = SftpGranule;
-module.exports.getGranuleId = getGranuleId;
-module.exports.getCmrFiles = getCmrFiles;
 module.exports.getRenamedS3File = getRenamedS3File;
 module.exports.copyGranuleFile = copyGranuleFile;
 module.exports.isFileRenamed = isFileRenamed;
