+++ packages/ingest/test/test-granule-extract-cmr.js
@@ -1,0 +1,18 @@
+const test = require('ava');
+
+const {
+  getGranuleId
+} = require('../granule-extract-cmr');
+
+test('getGranuleId is successful', (t) => {
+  const uri = 'test.txt';
+  const regex = '(.*).txt';
+  t.is(getGranuleId(uri, regex), 'test');
+});
+
+test('getGranuleId fails', (t) => {
+  const uri = 'test.txt';
+  const regex = '(.*).TXT';
+  const error = t.throws(() => getGranuleId(uri, regex), Error);
+  t.is(error.message, `Could not determine granule id of ${uri} using ${regex}`);
+});
